requirements.txt
PyMuPDF
chromadb
sentence-transformers
google-genai
streamlit